LRCI-7575 Localize mock Environment harness to its consumer test
